+++ tests/all_tests.rs
@@ -6,6 +6,7 @@ pub mod analysis_report_tests;
 pub mod analyzer_tests;
 pub mod args_tests;
 pub mod definition_analyzer_tests;
+pub mod humble_adapter_detector_tests;
 pub mod manifest_resolver_tests;
 pub mod report_printer_tests;
 pub mod runner_tests;